Wedding Halls/eventhalls/requirements.txt
asgiref==3.2.10
Babel==2.8.0
Django==3.1
django-filter==2.3.0
django-messages==0.6.0
django-phonenumber-field==4.0.0
phonenumbers==8.12.7
Pillow==7.2.0
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1